requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
pandas == 1.0.3
pandastable == 0.12.2.post1
psycopg2_binary == 2.8.5
requests == 2.23.0

